14315: Solve PR Remarks.
