+++ src/xcode/ENA/ENA/Source/Services/PPAccessControl/Model/ErrorCodeProviding.swift
@@ -1,0 +1,9 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+protocol ErrorCodeProviding: Error {
+	typealias ErrorCode = String
+	/// Error Code
+	var description: ErrorCode { get }
+}
