+++ src/xcode/ENA/ENA/Source/Services/PPAccessControl/Model/SRSError.swift
@@ -2,20 +2,6 @@
 // 🦠 Corona-Warn-App
 //
 
-import Foundation
-
-protocol SRSErrorAlertProviding: Error {
-	/// The SRS Error Alert type aims to share error message based on error codes.
-	/// The type is optional, because there are error cases of error enums, that hasn't a relationship to SRS
-	var srsErrorAlert: SRSErrorAlert? { get }
-}
-
-protocol ErrorCodeProviding: Error {
-	typealias ErrorCode = String
-	/// Error Code
-	var description: ErrorCode { get }
-}
-
 enum SRSError: Error, Equatable {
 	case ppacError(PPACError)
 	case otpError(OTPError)
