+++ src/xcode/ENA/ENA/Source/Services/PPAccessControl/Model/SRSErrorAlertProviding.swift
@@ -1,0 +1,9 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+protocol SRSErrorAlertProviding: Error {
+	/// The SRS Error Alert type aims to share error message based on error codes.
+	/// The type is optional, because there are error cases of error enums, that hasn't a relationship to SRS
+	var srsErrorAlert: SRSErrorAlert? { get }
+}
